requirements.txt
asttokens==2.4.1
bidict==0.23.1
blinker==1.7.0
click==8.1.7
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==1.26.4
nvidia-riva-client==2.15.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
Pygments==2.17.2
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
wcwidth==0.2.13
Werkzeug==3.0.2
wsproto==1.2.0
